Added CMake build system; restructure code separating headers, source and examples.
