+++ examples/contract_example.cpp
@@ -1,6 +1,9 @@
 #include <iostream>
 #include <string.h>
-#include "univcont.h"
+#include "uccontract.h"
+#include "ucontainer.h"
+#include "ucio.h"
+
 using namespace JAD;
 
 int main(int argc, char** argv)
@@ -9,10 +12,10 @@ int main(int argc, char** argv)
     UniversalContainer cspec = uc_from_json_file("contract.json");
     UCContract contract(cspec);
     UniversalContainer uc = uc_from_json_file("data.json");
-    
+
     contract.compare_and_throw(uc);
-    cout << "Data matches contract." << endl;
-    
+    std::cout << "Data matches contract." << std::endl;
+
     return 0;
   } catch(UniversalContainer uce) {
     print(uce);
