+++ include/string_util.h
