+++ version.cmake
@@ -1,0 +1,3 @@
+SET(APPLICATION_VERSION_MAJOR 0)
+SET(APPLICATION_VERSION_MINOR 1)
+SET(APPLICATION_VERSION_PATCH 0)
